pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "jupyter>=1.1.1",
    "kaleido>=1.2.0",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "plotly>=6.5.1",
    "python-dotenv>=1.2.1",
    "pytrends>=4.9.2",
    "requests>=2.32.5",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "selenium>=4.39.0",
    "supabase>=2.27.2",
    "tqdm>=4.67.1",
]
